// client/src/app/models/User.ts
export interface User{
    id: string;
    tipo_user: string;
    nombre: string;
    apellido: string;
    clave: string;
    email: string;
    telefono: string;
    fotografia: string;
    genero: string;
    fecha: string;
    direccion: string;
    estado: string;
    credito: string;
    ganancia: string;
    clase: string;
}